Update find command to accept Predicate<Person>
The find command originally was written to search a person by name.
However, with the new decision to accept 'name' as a common attribute
and 'phone' and 'email' as unique identifiers, and potentially more
unique identifiers in the future, the find command should not only
accept a predicate that matches a person's name, but other attributes
too. Or even, a predicate that can match multiple attributes at the same
time!

Let's loosen the FindCommand class to accept a general Predicate related
to a Person. In this iteration, only phone number and email are checked,
so two classes EmailIsKeywordPredicate and PhoneIsKeywordPredicate are
implemented, which exactly matches a person's email address and phone
number respectively.

Let's also add unit test cases for the new predicates and update the
test cases for the FindCommand class.